collab: Fix writing LLM rate limit events to Clickhouse
